[ai-assisted] refactor(types): rename ag-grid type path
Why:
- Address issue #60 by correcting the misspelled AG Grid shared type path.

What:
- Renamed src/types/ag-gird to src/types/ag-grid.
- Updated all imports and references from ag-gird to ag-grid.
- Updated CHANGELOG.md and the React maintainability plan.

Validation:
- rg 'ag-gird': no remaining references
- npm run typecheck: pass
- npm run lint: pass with existing warnings only
