+++ src/stores/studio/mgmt/acl.entries.store.ts
@@ -1,5 +1,5 @@
 import { createAclEntry, deleteAclEntry } from "@/data/studio/mgmt/acl";
-import type { PageableDataSource } from "@/types/ag-gird";
+import type { PageableDataSource } from "@/types/ag-grid";
 import type { AclEntryDto, AclEntryRequest } from "@/types/studio/acl";
 import { defineStore } from "pinia";
 import { AbstractPageDataSource } from "@/data/datasource/abstract.page.datasource";
