+++ src/stores/studio/mgmt/mail.inbox.store.ts
@@ -1,5 +1,5 @@
 import { AbstractPageDataSource } from "@/data/datasource/abstract.page.datasource";
-import type { PageableDataSource } from "@/types/ag-gird";
+import type { PageableDataSource } from "@/types/ag-grid";
 import type { MailMessageDto } from "@/types/studio/mail";
 import { api } from "@/data/http";
 import { defineStore } from "pinia";
