+++ src/stores/studio/mgmt/mail.synclog.store.ts
@@ -1,5 +1,5 @@
 import { AbstractPageDataSource } from "@/data/datasource/abstract.page.datasource";
-import type { PageableDataSource } from "@/types/ag-gird";
+import type { PageableDataSource } from "@/types/ag-grid";
 import { defineStore } from "pinia";
 
 type IPageableMailSyncLogDataSource = PageableDataSource & {
