+++ src/stores/studio/mgmt/roles.store.ts
@@ -1,5 +1,5 @@
 import { AbstractPageDataSource } from "@/data/datasource/abstract.page.datasource";
-import type { CreateOptions, PageableDataSource } from "@/types/ag-gird";
+import type { CreateOptions, PageableDataSource } from "@/types/ag-grid";
 import { api } from "@/data/http";
 import { defineStore } from "pinia";
